+++ app/models/payment.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+# Payment Model
+class Payment < ApplicationRecord
+  belongs_to :user
+end