+++ db/migrate/20230220001446_create_payments.rb
@@ -1,0 +1,10 @@
+class CreatePayments < ActiveRecord::Migration[7.0]
+  def change
+    create_table :payments  do |t|
+      t.string :paypal_id
+      t.references :user
+      t.json :details
+      t.timestamps
+    end
+  end
+end